Ajustes em Var para suporte a DateTime

+++ src/main/java/cronapi/math/Operations.java
@@ -28,7 +28,7 @@ public class Operations {
 			result = multiplyDouble(values);
 			break;
 		}
-		case LONG: {
+		case INT: {
 			result = multiplyLong(values);
 			break;
 		}
